Add basic Open Telemetry integration
This change also pulls in Prometheus and Ollama as dev-dependencies -
follow up commit to pull them out of the "prod" compose. This allows us
to run a full stack locally including a model which can be added to
liteLLM to test that we are pulling the correct telemetry from the
callbacks.