Add passthrough UI for OpenMeter and update grafana
OpenMeter doesn't seem to have its own UI in the self-hoste version
(that I have found) so rather than play with all the possible APIs I
have hacked on a "passthrough" in the amazee.ai admin interface allowing
easy testing of the various APIs.

Additionally we add some new graphs to the default grafana dashboard to
allow a user to understand what is happening. Data can be populated by
running the `random_llm` script in a loop.