Ensure alembic versions are recorded on startup
This is a brittle change, being put in place to future proof the
database migration system. There are already environments set up with
tables created from the models. This was done _after_ migrations had
been created, meaning if future migrations are attempted the first
revision always fails.

Future-proofing is done in two ways.
1. We assume that the models and migrations are in sync, and so when new
   environments are initialised we simply stamp the revision number in
   alaembic and allow that future changes will happen via migrations.
1. For existing setups we assume that there have been no database
   changes between the previous deployment and this commit, and stamp
   the latest revision if _any migration fails_. This is the brittle
   part, but still ensures that future migrations can occur safely.